rename cards edit css for description card elements
